+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed #162: Local stack not immediately updated if value changed from UI.
 - Fixed #106: When loading a 128K .sna or .z80 file the port 7FFD is now set correctly (according file) for CSpect and dezogif (ZXNext).
 - Discussion/Issue #165/#164: New LOGPOINT variables: ${Remote.tStates}, ${Remote.cpuFrequency} and ${Remote.slots}
+- Panel OUTPUT logging of "DeZog zsim: Custom Code" and "DeZog zsim: Hardware" have been merged. New name is "DeZog zsim".
 
 # 3.6.5
 - Debug command "-sjasmplus.path" added for dynamic loading of sld files during runtime (#156).